Moving attachment methods out of utils
